+++ src/main/java/com/percent99/OutSpecs/repository/ReactionRepository.java
@@ -5,7 +5,11 @@ import com.percent99.OutSpecs.entity.ReactionType;
 import com.percent99.OutSpecs.entity.TargetType;
 import com.percent99.OutSpecs.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 @Repository
 public interface ReactionRepository extends JpaRepository<Reaction, Long> {
@@ -25,4 +29,20 @@ public interface ReactionRepository extends JpaRepository<Reaction, Long> {
      */
     long countByTargetTypeAndTargetIdAndReactionType(TargetType targetType, Long targetId, ReactionType reactionType);
 
+    /**
+     * User가 follow한 targetId 목록 찾기
+     * @param user
+     * @return User가 follow한 targetId 목록
+     */
+    @Query("SELECT r.targetId FROM Reaction r WHERE r.user = :user AND r.reactionType = 'FOLLOW' AND r.targetType = 'USER'")
+    List<Long> findFollowedUserIds(@Param("user") User user);
+
+
+    /**
+     * User가 북마크한 Post targetId 목록 찾기
+     * @param user
+     * @return User가 북마크한 Post targetId 목록
+     */
+    @Query("SELECT r.targetId FROM Reaction r WHERE r.user = :user AND r.reactionType = 'BOOKMARK' AND r.targetType = 'POST'")
+    List<Long> findBookmarkedPostIdsByUser(@Param("user") User user);
 }
