--- src/main/java/com/percent99/OutSpecs/service/Impl/UserServiceImpl.java
@@ -1,112 +1,0 @@
-package com.percent99.OutSpecs.service.Impl;
-
-import com.percent99.OutSpecs.entity.User;
-import com.percent99.OutSpecs.entity.UserDTO;
-import com.percent99.OutSpecs.repository.ProfileRepository;
-import com.percent99.OutSpecs.repository.UserRepository;
-import com.percent99.OutSpecs.service.UserService;
-import jakarta.persistence.EntityExistsException;
-import jakarta.persistence.EntityNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-/**
- *  User 정보 등록, 수정, 삭제, 조회 등을 위한 서비스
- */
-@Service
-public class UserServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-    private final ProfileRepository profileRepository;
-
-    public UserServiceImpl(UserRepository userRepository, ProfileRepository profileRepository) {
-        this.userRepository = userRepository;
-        this.profileRepository = profileRepository;
-    }
-
-    /**
-     * 회원가입 정보 저장
-     * @param userDTO
-     * @return 저장된 user
-     */
-    @Override
-    public User registerUser(UserDTO userDTO) {
-        if (userRepository.existsByUsername(userDTO.getUsername())) {
-            throw new EntityExistsException("이미 존재하는 회원입니다.");
-        }
-        if(userDTO.getPassword() != null && !userDTO.getPassword().isEmpty()) {
-           if(userDTO.getPassword().length() <= 2) {
-               throw new IllegalArgumentException("비밀번호는 3자리 이상이어야 합니다.");
-           }
-        }
-
-        User user = new User();
-        user.setUsername(userDTO.getUsername());
-        user.setPassword(userDTO.getPassword());
-        user.setRole(userDTO.getRole());
-        user.setAiRateLimit(userDTO.getAiRateLimit());
-
-        if(userDTO.getProviderId() != null) {
-            user.setProviderId(userDTO.getProviderId());
-        }
-        return userRepository.save(user);
-    }
-
-    /**
-     * 유저정보 조회
-     * @param username
-     * @return username으로 찾은 유저정보
-     */
-    @Override
-    public Optional<User> findByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
-
-    /**
-     * 유저정보 수정
-     * @param userDTO
-     * @return 수정된 유저정보
-     */
-    @Override
-    public User updateUser(UserDTO userDTO) {
-        User exiting = userRepository.findByUsername(userDTO.getUsername())
-                .orElseThrow(() -> new EntityNotFoundException("해당 유저는 존재하지 않습니다."));
-
-        exiting.setRole(userDTO.getRole());
-        if(exiting.getProviderId() == null &&
-                userDTO.getPassword() != null && !userDTO.getPassword().isEmpty()) {
-            if(userDTO.getPassword().length() <= 2) {
-                throw new IllegalArgumentException("비밀번호는 3자리 이상이어야 합니다.");
-            }
-            exiting.setPassword(userDTO.getPassword());
-        }
-
-        return userRepository.save(exiting);
-    }
-
-    /**
-     * AI 사용횟수 감소
-     * @param userId
-     */
-    @Override
-    public void decrementAiRateLimit(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 유저는 존재하지 않습니다."));
-        user.setAiRateLimit(user.getAiRateLimit() - 1);
-        userRepository.save(user);
-    }
-
-    /**
-     * 유저정보 삭제
-     * @param userId
-     */
-    @Override
-    public void deleteUserAndProfile(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityNotFoundException("해당 유저는 존재하지 않습니다."));
-
-        profileRepository.deleteByUserId(userId);
-        userRepository.delete(user);
-    }
-}
